requirements.txt
fastapi>=0.100.0
uvicorn[standard]
torch==2.3.0
torchvision==0.18.0
transformers==4.43.0
Pillow==10.3.0
PyPDF2==3.0.1
accelerate==0.30.0
pydantic>=2.0.0
python-multipart==0.0.9